feat: implement data saving strategies and integrate into nested sampling
